pyproject.toml
[tool.poetry]
name = "kurs-3-"
version = "0.1.0"
description = ""
authors = ["ow3rr <lighteckiy@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
